#218 Implement subscribe/unsubscribe to Schedule Events. Fix tests

+++ src/main/java/io/github/externschool/planner/service/ScheduleServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.DayOfWeek;
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -27,6 +28,8 @@ import java.time.temporal.WeekFields;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.ReentrantLock;
 import java.util.stream.Collectors;
 
 import static io.github.externschool.planner.util.Constants.LOCALE;
@@ -62,7 +65,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     public ScheduleEvent createEvent(User owner, ScheduleEventReq eventReq) {
         ScheduleEventType type = this.eventTypeRepository.findByName(eventReq.getEventType());
 
-        canUserHandleEventForType(owner, type);
+        canUserOwnAnEventForType(owner, type);
 
         ScheduleEvent newEvent = ScheduleEvent.builder()
                 .withTitle(eventReq.getTitle())
@@ -81,7 +84,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     public ScheduleEvent createEventWithDuration(final User owner, final ScheduleEventDTO eventDTO, final int minutes) {
 
         ScheduleEventType type = eventTypeRepository.findByName(eventDTO.getEventType());
-        canUserHandleEventForType(owner, type);
+        canUserOwnAnEventForType(owner, type);
 
         ScheduleEvent newEvent = ScheduleEvent.builder()
                 .withTitle(eventDTO.getTitle())
@@ -121,17 +124,18 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public List<ScheduleEvent> getEventsByOwner(User owner) {
-        return eventRepository.findAllByOwner(owner);
+    public List<ScheduleEvent> getEventsByOwnerStartingBetweenDates(final User owner,
+                                                                    final LocalDate firstDate,
+                                                                    final LocalDate lastDate) {
+        return eventRepository.findAllByOwnerAndStartOfEventBetweenOrderByStartOfEvent(
+                owner,
+                firstDate.atStartOfDay(),
+                lastDate.atTime(LocalTime.MAX));
     }
 
     @Override
-    public void cancelEventById(long id) {
-        eventRepository.findById(id).ifPresent(event -> {
-            event.setCancelled(true);
-            event.setOpen(false);
-            event.setModifiedAt(LocalDateTime.now());
-        });
+    public List<ScheduleEvent> getEventsByOwner(User owner) {
+        return eventRepository.findAllByOwner(owner);
     }
 
     @Override
@@ -139,6 +143,25 @@ public class ScheduleServiceImpl implements ScheduleService {
         ScheduleEventType eventType = eventTypeRepository.findByName(type.getName());
 
         return eventRepository.findAllByType(eventType);
+    }
+
+    @Override
+    public void cancelEventByIdAndSave(long id) {
+        eventRepository.findById(id).ifPresent(event -> {
+            event.setCancelled(true);
+            event.setOpen(false);
+            event.setModifiedAt(LocalDateTime.now());
+            eventRepository.save(event);
+        });
+    }
+
+    @Override
+    public void findEventByIdSetOpenAndSave(final long id, final boolean state) {
+        eventRepository.findById(id).ifPresent(event -> {
+            event.setOpen(state);
+            event.setModifiedAt(LocalDateTime.now());
+            eventRepository.save(event);
+        });
     }
 
     @Override
@@ -157,7 +180,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Override
     public ScheduleEvent addOwner(final User owner, final ScheduleEvent event) {
         if (owner != null && event != null && event.getOwner() == null) {
-            canUserHandleEventForType(owner, event.getType());
+            canUserOwnAnEventForType(owner, event.getType());
 
             owner.addOwnEvent(event);
             eventRepository.save(event);
@@ -176,45 +199,68 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Override
     public Optional<Participant> addParticipant(User user, ScheduleEvent event) {
         Optional<Participant> optionalParticipant = Optional.empty();
-        // TODO check for user rights to participate in this event
-        // TODO set number of users by event type
-        if (user != null && event != null && event.isOpen()) {
-            Participant participant = new Participant(user, event);
-            user.addParticipant(participant);
-            event.addParticipant(participant);
-            event.setModifiedAt(LocalDateTime.now());
-            optionalParticipant = Optional.of(participant);
+        if (user != null
+                && event != null
+                && event.isOpen()) {
+            canUserParticipateInEventForType(user, event.getType());
+            int maximum = event.getType().getAmountOfParticipants();
 
-            participantRepository.save(participant);
-            eventRepository.save(event);
+            ReentrantLock lock = new ReentrantLock();
+            try {
+                if (lock.tryLock(5, TimeUnit.SECONDS) && event.getParticipants().size() < maximum) {
+                    Participant participant = new Participant(user, event);
+                    if (event.getParticipants().size() >= maximum) {
+                        event.setOpen(false);
+                    }
+                    event.setModifiedAt(LocalDateTime.now());
+                    optionalParticipant = Optional.ofNullable(participantRepository.save(participant));
+                    eventRepository.save(event);
+                }
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            } finally {
+                lock.unlock();
+            }
         }
 
         return optionalParticipant;
     }
 
     @Override
-    public Optional<Participant> getParticipantByUserAndEvent(final User user, final ScheduleEvent event) {
+    public Optional<Participant> findParticipantByUserAndEvent(final User user, final ScheduleEvent event) {
         return participantRepository.findParticipantByUserAndEvent(user, event);
     }
 
     @Override
     public void removeParticipant(final Participant participant) {
-        participantRepository.findById(participant.getId()).ifPresent(p -> {
-            Optional.ofNullable(participant.getUser()).ifPresent(user -> {
-                user.removeParticipant(participant);
-                participantRepository.save(participant);
-            });
-            Optional.ofNullable(participant.getEvent()).ifPresent(event -> {
-                event.removeParticipant(participant);
-                participantRepository.save(participant);
-            });
-            participantRepository.delete(participant);
-        });
+        ReentrantLock lock = new ReentrantLock();
+        try {
+            if (lock.tryLock(5, TimeUnit.SECONDS)) {
+                participantRepository.findById(participant.getId()).ifPresent(p -> {
+                    Optional.ofNullable(participant.getUser()).ifPresent(user -> {
+                        user.removeParticipant(participant);
+                        participantRepository.save(participant);
+                    });
+                    Optional.ofNullable(participant.getEvent()).ifPresent(event -> {
+                        event.removeParticipant(participant);
+                        participantRepository.save(participant);
+                    });
+                    participantRepository.delete(participant);
+                });
+            }
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } finally {
+            lock.unlock();
+        }
     }
 
     @Override
     public LocalDate getCurrentWeekFirstDay() {
         LocalDate now = LocalDate.now();
+        if (now.getDayOfWeek() == DayOfWeek.SATURDAY || now.getDayOfWeek() == DayOfWeek.SUNDAY) {
+            now = now.plusDays(2L);
+        }
         TemporalField fieldISO = WeekFields.of(LOCALE).dayOfWeek();
 
         return now.with(fieldISO, 1);
@@ -235,8 +281,7 @@ public class ScheduleServiceImpl implements ScheduleService {
         return getCurrentWeekFirstDay().plus(Period.of(0, 0, 7));
     }
 
-    private void canUserHandleEventForType(User user, ScheduleEventType type) {
-
+    private void canUserOwnAnEventForType(User user, ScheduleEventType type) {
         if (user != null && type != null) {
             for (Role role : user.getRoles()) {
                 if (type.getOwners().contains(role)) {
@@ -244,9 +289,22 @@ public class ScheduleServiceImpl implements ScheduleService {
                 }
             }
         }
-
         throw new UserCannotHandleEventException(
                 String.format("The user %s is not allowed to create or own %s type of event",
+                        user != null ? user.getEmail() : "NULL",
+                        type != null ? type.getName() : "NULL"));
+    }
+
+    private void canUserParticipateInEventForType(User user, ScheduleEventType type) {
+        if (user != null && type != null) {
+            for (Role role : user.getRoles()) {
+                if (type.getParticipants().contains(role)) {
+                    return;
+                }
+            }
+        }
+        throw new UserCannotHandleEventException(
+                String.format("The user %s is not allowed to participate in %s type of event",
                         user != null ? user.getEmail() : "NULL",
                         type != null ? type.getName() : "NULL"));
     }
